% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_bin.R
\name{env_bin}
\alias{env_bin}
\title{Create temporally binned predictor layer according to species traits}
\usage{
env_bin(
  sp_trait,
  env_path,
  out_dir,
  var_name,
  lon_min,
  lon_max,
  lat_min,
  lat_max,
  date_min,
  date_max,
  lag_days,
  temporal_bin
)
}
\arguments{
\item{sp_trait}{Character. Species vertical habitat use. One of \code{"pelagic"},
\code{"benthic"}, or \code{"benthopelagic"}.}

\item{env_path}{Character. Path to the downloaded CMEMS NetCDF file.}

\item{out_dir}{Character. Directory where the temporally‑binned raster layers
will be saved.}

\item{var_name}{Character. Name of the variable (refer to CMEMS user manual).}

\item{lon_min}{Numeric. Minimum longitude (western boundary).}

\item{lon_max}{Numeric. Maximum longitude (eastern boundary).}

\item{lat_min}{Numeric. Minimum latitude (southern boundary).}

\item{lat_max}{Numeric. Maximum latitude (northern boundary).}

\item{date_min}{Date (or character convertible to Date). Start date.}

\item{date_max}{Date (or character convertible to Date). End date.}

\item{lag_days}{Numeric. Number of days ocean seasonality lags behind
atmospheric seasonality. e.g., \code{76} if missing.}

\item{temporal_bin}{Character. How a year will be split. One of \code{"monthly"},
\code{"quarterly"}, or \code{"seasonal"}.}
}
\value{
No return value. The function writes raster files (ASCII format) to
subdirectories of \code{out_dir} (e.g., \code{out_dir/Jan/thetao_mean.asc},
\code{out_dir/Q1/thetao_mean.asc}, \code{out_dir/Spring/thetao_mean.asc}).
}
\description{
Create temporally binned predictor layer according to species traits
}
\examples{
\donttest{
## sea surface height
env_bin(sp_trait="pelagic",
       env_path="F:/cmems/IndoWPac_slh_1m_y2015_y2024.nc",
       out_dir="F:/whaleshark_sdm/bin",
       var_name="zos_mean",
       lon_min=90,
       lon_max=180,
       lat_min=-41,
       lat_max=40,
       date_min=ymd(20150101),
       date_max = ymd(20241201),
       temporal_bin="quarterly")
       }
}
