% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_download.R
\name{env_download}
\alias{env_download}
\title{Download environmental layers from CMEMS}
\usage{
env_download(
  path_cmems_tool,
  out_dir,
  productId,
  variable,
  date_min,
  date_max,
  lon,
  lat,
  depth,
  out_name
)
}
\arguments{
\item{path_cmems_tool}{Character. Path to the copernicusmarine executable,
e.g. \code{"/miniconda/envs/copernicusmarine/Scripts/copernicusmarine.exe"}.}

\item{out_dir}{Character. Directory where output data will be stored.}

\item{productId}{Character. ID of the CMEMS product (see dataset user manual
in the CMEMS product webpage).
e.g., \code{"cmems_mod_glo_phy-mnstd_my_0.25deg_P1M-m"}.}

\item{variable}{Character. Variable name to download. e.g., "thetao_mean".}

\item{date_min}{Date or character. Minimum date. e.g., "1993-01-01".}

\item{date_max}{Date or character. Maximum date. e.g., "2022-12-01".}

\item{lon}{Numeric vector of length 2. Longitude range (min, max).
e.g., \code{c(-180, 179.75)}.}

\item{lat}{Numeric vector of length 2. Latitude range (min, max).
e.g., \code{c(-80, 90)}.}

\item{depth}{Numeric vector of length 2. Depth range in meters (min, max).
e.g., \code{c(0.51, 199.79)}.}

\item{out_name}{Character. Output file name ending with \code{.nc}.}
}
\value{
No return value. The function downloads a NetCDF file to \code{out_dir}.
}
\description{
Download environmental layers from CMEMS
}
\examples{
\donttest{
path_cms<-"C:/shark/copernicusmarine.exe" # replace to your own directory
out_dir <- "F:/whaleshark_sdm/cmems" # replace to your own directory
productId <- "cmems_mod_glo_phy-mnstd_my_0.25deg_P1M-m"
variable <- " --variable thetao_mean"
date_min <- lubridate::ymd(20110101)
date_max <- lubridate::ymd(20120101)
lon <- c(100, 130)
lat <- c(15, 30)
depth <- c(0.51, 50)
out_name <- "Pacific_temp_05_50_2011.nc"
env_download(path_cms,out_dir,productId,variable,
            date_min,date_max,lon,lat,
            depth,out_name)
}
}
