% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_temp.R
\name{fit_temp}
\alias{fit_temp}
\title{Title This function was used to fit temperature-bin data from satellite
studies on oceanic sharks}
\usage{
fit_temp(ws, factor)
}
\arguments{
\item{ws}{the temperature-bin data with three columns: freq,lower,upper.
freq, the frequency of occurrences; lower, the lower bound of the
temperature bin; upper, the upper bound of the temperature bin.
See an example data "ws.RData" in the folder "data".}

\item{factor}{the conversion factor that was used to transform original
temperature to a value between 0 and 1.}
}
\value{
This function returns the model fit, AIC values, the estimated
preferred and tolerable temperatures based on the best fitted model，
and a ggplot of the model fits.
}
\description{
Title This function was used to fit temperature-bin data from satellite
studies on oceanic sharks
}
\examples{
data(ws)
fit_temp(ws,30)
}
