% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB_depth.R
\name{TB_depth}
\alias{TB_depth}
\title{Calculate depth range within specified temperature range for each grid cell
using adaptive interpolation based on temp_bin results}
\usage{
TB_depth(
  temp_bin_path,
  PTR,
  time_type = c("month", "quarter", "season"),
  layer_option = "top",
  save_path,
  na_value = -9999,
  use_makima = TRUE,
  N_cells = 50,
  ncores = 1
)
}
\arguments{
\item{temp_bin_path}{Character. Root path of temp_bin generated files.}

\item{PTR}{Numeric vector of length 2. Temperature range, e.g., \code{c(23.5, 30.2)}.}

\item{time_type}{Character. Time type, should match that used in temp_bin:
\code{"month"}, \code{"quarter"}, or \code{"season"}.}

\item{layer_option}{Character. Which depth interval to extract: \code{"top"} (shallowest)
or \code{"bottom"} (deepest). Default is \code{"top"}.}

\item{save_path}{Character. Path to save output files.}

\item{na_value}{Numeric. Value to replace NA (default: -9999).}

\item{use_makima}{Logical. Whether to use makima interpolation. If \code{FALSE} , falls back to spline.}

\item{N_cells}{Integer. Number of random cells to plot (5--30). If \code{0}, no plot is generated.}

\item{ncores}{Integer. Number of cores for parallel processing (uses \code{pbapply}). Default: 1.}
}
\value{
Invisibly returns a character vector of all generated file paths.
}
\description{
Calculate depth range within specified temperature range for each grid cell
using adaptive interpolation based on temp_bin results
}
\examples{
\donttest{
df<-read.csv(system.file("extdata","summary_niche_fit.csv",package = "OceanSDM"))
PTR<-df[2:3,2] # preferred temperature range
TTR<-df[4:5,2] # tolerable temperature range
temp_bin_path<-"F:/whaleshark_sdm/4D" # the same path as the save_path of the clim_bin()
save_path_TPD<-"F:/whaleshark_sdm/4D/TPD" ## new path for storing data of Preferred Depth (PD)
save_path_TTD<-"F:/whaleshark_sdm/4D/TTD" ## new path for storing data of Tolerable Depth (TD)
TTD_quarter<-TB_depth(temp_bin_path, TTR, time_type = "quarter",
                     layer_option="top",
                     save_path_TTD, na_value = -9999,
                        use_makima = TRUE,
                         N_cells = 50)
}
}
