% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB_sdm.R
\name{TB_sdm}
\alias{TB_sdm}
\title{Build Species Distribution Model using \code{sdm} and \code{blockCV} for spatial cross-validation}
\usage{
TB_sdm(
  env_path,
  occ_path,
  res_path,
  species,
  x,
  y,
  pre_abs_ratio = 1,
  collinearity_thresh = 0.7,
  remove_collinear = FALSE,
  block_size = NULL,
  algorithms = c("glm", "gam", "brt", "rf", "svm"),
  cv_method = "random",
  n_folds = 5,
  var.importance = TRUE,
  response.curve = TRUE,
  ensemble.method = "weighted",
  ensemble.stat = "AUC",
  expr = "auc > 0.8 & tss > 0.6",
  ncore = 1,
  seed = 123,
  proj_path = NULL
)
}
\arguments{
\item{env_path}{character. Directory containing environmental rasters (.asc or .tif).}

\item{occ_path}{character. Path to occurrence file (tab‑delimited).}

\item{res_path}{character. Output directory.}

\item{species}{character. Column name for species (not used, kept for compatibility).}

\item{x}{character. Column name for longitude.}

\item{y}{character. Column name for latitude.}

\item{pre_abs_ratio}{numeric. Ratio of presences to pseudo-absences points.}

\item{collinearity_thresh}{numeric. Correlation threshold.}

\item{remove_collinear}{logical. Remove correlated variables.}

\item{block_size}{numeric. Block size (unit: meters) can be user defined if you did not want to find it through the time-consuming auto-process based on Moran's I curve}

\item{algorithms}{character vector. sdm methods.}

\item{cv_method}{character. type of assignment of spatial blocks into folds. Can be random (default), systematic, or checkerboard.}

\item{n_folds}{integer. Number of folds for spatial cross-validation.}

\item{var.importance}{logical.}

\item{response.curve}{logical.}

\item{ensemble.method}{character. "weighted" or "unweighted".}

\item{ensemble.stat}{character. "AUC" or "TSS".}

\item{expr}{character. Select which models will be used for ensemble based on AUC and/or TSS, e.g. expr = 'auc > 0.8 & tss > 0.6'.}

\item{ncore}{integer. Number of cores.}

\item{seed}{integer. Random seed.}

\item{proj_path}{character. Optional projection rasters directory.}

\item{max_dist}{numeric. Maximum distance (unit: meters) to be used in calculating Moran's I. Default is NULL}
}
\value{
A list containing:
\item{final_model}{A final sdm model object}
\item{ensemble}{Raster object of the predicted habitat suitability by the ensemble model}
\item{evaluation}{Model evaluation table}
\item{var_imp}{Variable importance table}
\item{response_curves}{Data frame of response curve data (from sdm::rcurve)}
\item{optimal_block}{Identified optimal block size}
\item{collinearity_report}{Collinearity analysis report}
}
\description{
Build Species Distribution Model using \code{sdm} and \code{blockCV} for spatial cross-validation
}
\examples{
\donttest{
d_occ <- system.file("extdata", "R_typus_q1.txt", package = "OceanSDM")
d_env <- system.file("extdata", "Q1", package = "OceanSDM")
d_prj<-system.file("extdata", "proj", package = "OceanSDM")
d_res <- "F:/whaleshark_sdm/result_sdm/Q1" # Change to your own directory
library(sdm) # initializing the internal state of 'sdm' package
sdm_q1 <- TB_sdm(
  env_path = d_env,
  occ_path = d_occ,
  res_path = d_res,
  species = "Species",
  x = "x",
  y = "y",
  pre_abs_ratio = 1,
  collinearity_thresh = 0.7,
  remove_collinear = TRUE,
  block_size = 500000,
  algorithms = c("brt", "rf", "svm"),
  cv_method = "random",
  n_folds = 5,
  var.importance = TRUE,
  response.curve = TRUE,
  ensemble.method = "weighted",
  ensemble.stat = "TSS",
  ncore = 1,
  seed = 123,
  proj_path = d_prj)
}
}
