% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ws}
\alias{ws}
\title{The example temperature data}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{class}{character, 10 levels of temperature bins}
\item{freq}{integer, from 29 to 4433}
\item{lower}{float from 0.1 to 28.1}
\item{upper}{float from 5 to 30}
...
}
}
\usage{
ws
}
\description{
A dataset containing temperature bins for whale sharks
}
\keyword{datasets}
