% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_check.R
\name{bias_check}
\alias{bias_check}
\title{Check spatial and temporal sampling bias}
\usage{
bias_check(
  occ_data,
  temporal_level,
  lag_days,
  spatial_method,
  world_path,
  range_path,
  prj = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{occ_data}{Data frame. Species occurrence records. Must contain columns:
\code{Species}, \code{x}, \code{y}, \code{year}, \code{month}, \code{day}.}

\item{temporal_level}{Character. Temporal level at which temporal sampling
bias is to be tested. One of \code{month}, \code{quarter}, or \code{season}.}

\item{lag_days}{Numeric. Number of days ocean seasonality lags behind.}

\item{spatial_method}{Character. Method used to generate species habitat area
for spatial random sampling. One of \code{convex_hull} or \code{range_map}.}

\item{world_path}{Character. File path to a world map shapefile (required when
\code{spatial_method = "convex_hull"}).}

\item{range_path}{Character. File path to a species range map shapefile
(required when \code{spatial_method = "range_map"}). Must contain a column
\code{sci_name}.}

\item{prj}{Character. Coordinate reference system to be used for spatial maps,
given as a PROJ string (e.g., "+proj=longlat +datum=WGS84").}
}
\value{
A list of class \code{BiasCheckResult} containing the following components:
\item{Temporal_level}{The temporal level used for testing.}
\item{Spatial_method}{The spatial method used.}
\item{Statistical_tests}{List of statistical test results for temporal
bias and for spatial bias per temporal group.}
\item{Plots}{List of \code{ggplot} objects showing temporal and spatial bias.}
\item{Summary}{List with summary statistics (e.g., total records,
number of temporal groups).}
}
\description{
Check spatial and temporal sampling bias
}
\examples{
\donttest{
library(dplyr)
## Read input data or skip this line if you have run the section 1.4
corrected_ob<-read.csv(system.file("extdata", "R_typus_adj.csv", package = "OceanSDM"))
## Read input data or skip this line if you have run the section 1.4
range_path <-system.file("extdata", "R_typus_wip2.shp", package = "OceanSDM")
colnames(corrected_ob)
occ_ob<-corrected_ob[,c(1:7)] # removed unuseful columns
colnames(occ_ob)[2:3]<-c("y","x") # rename latitude and longitude columns
#### For the quarterly SDMs
occ_ob2 <- occ_ob \%>\%
  mutate(
      quarter = dplyr::case_when(
            month \%in\% 1:3   ~ "Q1",
                  month \%in\% 4:6   ~ "Q2",
                     month \%in\% 7:9  ~ "Q3",
                       month \%in\% 10:12 ~ "Q4",
                         TRUE ~ NA_character_
                         ))\%>\% dplyr::filter(!is.na(quarter))
bias_test <- bias_check(occ_data=occ_ob,
                        temporal_level="quarter",
                        spatial_method="range_map",
                        range_path=range_path,
                        prj = "+proj=longlat +datum=WGS84")
}
}
