% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_bin.r
\name{clim_bin}
\alias{clim_bin}
\title{Calculate mean water temperature (or any variable) for each grid cell and depth layer
within user-specified time periods. For seasonal aggregation, hemisphere is
automatically determined by latitude (north: >=0, south: <0).}
\usage{
clim_bin(
  clim_nc,
  presence_asc,
  time_type = c("month", "quarter", "season"),
  var_name = "thetao_mean",
  save_path,
  na_value = -9999
)
}
\arguments{
\item{clim_nc}{Character. Path to the multi-depth layers over years
and months of climate NetCDF file.}

\item{presence_asc}{List or SpatRaster. A named list where each element is a
presence raster (values 1 = presence, NA/0 = absence) corresponding to a
temporal group (folder name). For season, names must be one of
"Winter", "Spring", "Summer", "Autumn".}

\item{time_type}{Character. Type of temporal aggregation. One of \code{"month"},
\code{"quarter"}, or \code{"season"}.}

\item{var_name}{Character. Variable name in NetCDF. If not found exactly,
the function will attempt to match (case‑insensitive) and warn.
e.g., \code{"thetao_mean"}.}

\item{save_path}{Character. Directory path where output files will be saved.}

\item{na_value}{Numeric. Value to replace \code{NA}. Default \code{-9999}.}
}
\value{
Invisibly returns a character vector of all generated file paths.
}
\description{
Calculate mean water temperature (or any variable) for each grid cell and depth layer
within user-specified time periods. For seasonal aggregation, hemisphere is
automatically determined by latitude (north: >=0, south: <0).
}
\examples{
\donttest{
quarter_files<-c(
  Q1=system.file("extdata","result_sdm/Q1/predicted_presence.asc",package = "OceanSDM"),
  Q2=system.file("extdata","result_sdm/Q2/predicted_presence.asc",package = "OceanSDM"),
  Q3=system.file("extdata","result_sdm/Q3/predicted_presence.asc",package = "OceanSDM"),
  Q4=system.file("extdata","result_sdm/Q4/predicted_presence.asc",package = "OceanSDM"))
 presence_asc <- lapply(quarter_files, terra::rast)
 save_path<-"F:/whaleshark_sdm/4D"
 temp_nc<-system.file("extdata","cms/IndoWPac_temp_10m_y2024.asc",package = "OceanSDM") #' Replace to your own directory for 0-200m
 clim_bin(temp_nc, presence_asc, time_type = "quarter",
         save_path=save_path, na_value = -9999)
 }
}
