% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_bin.R
\name{clim_bin}
\alias{clim_bin}
\title{Calculate mean water temperature (or any variable) for each grid cell and depth layer
within user-specified time periods. For seasonal aggregation, hemisphere is
automatically determined by latitude (north: >=0, south: <0).}
\usage{
clim_bin(
  clim_nc,
  presence_asc,
  time_type = c("month", "quarter", "season"),
  var_name = "thetao_mean",
  save_path,
  na_value = -9999
)
}
\arguments{
\item{clim_nc}{Character. Path to the multi-depth layers over years
and months of climate NetCDF file.}

\item{presence_asc}{List or SpatRaster. A named list where each element is a
presence raster (values 1 = presence, NA/0 = absence) corresponding to a
temporal group (folder name). For season, names must be one of
"Winter", "Spring", "Summer", "Autumn".}

\item{time_type}{Character. Type of temporal aggregation. One of \code{"month"},
\code{"quarter"}, or \code{"season"}.}

\item{var_name}{Character. Variable name in NetCDF. If not found exactly,
the function will attempt to match (case‑insensitive) and warn. Default is
\code{"thetao_mean"}.}

\item{save_path}{Character. Directory path where output files will be saved.}

\item{na_value}{Numeric. Value to replace \code{NA}. Default \code{-9999}.}
}
\value{
Invisibly returns a character vector of all generated file paths.
}
\description{
Calculate mean water temperature (or any variable) for each grid cell and depth layer
within user-specified time periods. For seasonal aggregation, hemisphere is
automatically determined by latitude (north: >=0, south: <0).
}
