% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_bin.r
\name{clim_bin}
\alias{clim_bin}
\title{Calculate mean water temperature for each grid cell and depth layer within user-specified time periods}
\usage{
clim_bin(
  clim_nc,
  presence_tif,
  time_type = c("month", "quarter", "season"),
  save_path,
  hemisphere = "north",
  na_value = -9999
)
}
\arguments{
\item{clim_nc}{Character. Path to the multi-depth layers over years and months of climate NetCDF file (e.g., ocean temperature).}

\item{presence_tif}{List or RasterLayer. A named list where each element is a
presence raster (values 1 = presence, NA/0 = absence) corresponding to a
temporal group (folder name). Alternatively, a single raster (not recommended
for multi‑group processing).}

\item{time_type}{Character. Type of temporal aggregation. One of \code{"month"}
(monthly mean), \code{"quarter"} (quarterly mean), or \code{"season"} (seasonal mean).}

\item{save_path}{Character. Directory path where output files (organized in
subfolders per temporal group) will be saved.}

\item{hemisphere}{Character. Hemisphere setting for seasonal aggregation.
Either \code{"north"} (Northern Hemisphere) or \code{"south"} (Southern Hemisphere).
Default is \code{"north"}.}

\item{na_value}{Numeric. Value used to replace \code{NA} in the output rasters.
Default is \code{-9999}.}
}
\value{
Invisibly returns a character vector of all generated file paths.
The function primarily writes ASCII raster files (\code{.asc}) into subfolders
under \code{save_path}.
}
\description{
Calculate mean water temperature for each grid cell and depth layer within user-specified time periods
}
\examples{
\donttest{
library(raster)
quarter_files<-c(
  Q1="F:/whaleshark_sdm/result_sdm/Q1/Ensemble/SUP/MAX_TSS/Rhincodon_typus.tif",
  Q2="F:/whaleshark_sdm/result_sdm/Q2/Ensemble/SUP/MAX_TSS/Rhincodon_typus.tif",
  Q3="F:/whaleshark_sdm/result_sdm/Q3/Ensemble/SUP/MAX_TSS/Rhincodon_typus.tif",
  Q4="F:/whaleshark_sdm/result_sdm/Q4/Ensemble/SUP/MAX_TSS/Rhincodon_typus.tif")
presence_tif <- lapply(quarter_files, raster)
save_path<-"F:/whaleshark_sdm/4D"
clim_nc<-"F:/whaleshark_sdm/cmems/IndoWPac_temp_200m_y2015_y2024.nc"
clim_bin(clim_nc, presence_tif, time_type = "quarter",
         save_path, hemisphere = "north", na_value = -9999)
}
}
