% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_ob.R
\name{correct_ob}
\alias{correct_ob}
\title{Correct occurrence coordinates based on a standard raster layer of the predictor dataset}
\usage{
correct_ob(
  occ_path,
  predictor_raster_path,
  max_distance_km = NULL,
  out_dir,
  verbose = TRUE,
  plot = FALSE,
  plot_file = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{occ_path}{Character. Path to occurrence CSV file. Must contain
columns \code{decimalLongitude} and \code{decimalLatitude} (WGS84 coordinates).}

\item{predictor_raster_path}{Character. File path to the standard raster layer
(e.g., an ASCII or GeoTIFF file) used as reference for correction.}

\item{max_distance_km}{Numeric. Maximum distance (in kilometres) that a point
will be adjusted to the nearest grid cell with non-NA value.}

\item{out_dir}{Character. \strong{Full file path} (including \code{.csv} extension) where
the corrected occurrence data will be saved. If the directory does not exist,
it will be created. If a directory path is given instead, a default filename
\code{corrected_occurrences.csv} will be appended.}

\item{verbose}{Logical. If \code{TRUE}, print progress messages. Default is \code{TRUE}.}

\item{plot}{Logical. If \code{TRUE}, generate a map showing correction status. Default \code{FALSE}.}

\item{plot_file}{Character. Optional file path to save the plot as TIFF (e.g., "plot.tif").
If \code{NULL}, the plot is displayed in the graphics device.}

\item{xlim}{Numeric vector of length 2 (min, max) for longitude limits. If \code{NULL},
automatically determined from the raster extent.}

\item{ylim}{Numeric vector of length 2 (min, max) for latitude limits. If \code{NULL},
automatically determined from the raster extent.}
}
\value{
A data frame (invisibly) with the corrected occurrences.
}
\description{
Correct occurrence coordinates based on a standard raster layer of the predictor dataset
}
\examples{
\donttest{
ev_layer<-system.file("extdata", "Q1/Depth.asc", package = "OceanSDM")
occ_range<-system.file("extdata", "R_typus_r.csv", package = "OceanSDM")
output_path<-"F:/whaleshark_sdm/occ/R_typus_adj.csv"
corrected_ob <- correct_ob(
  occ_path = occ_range,
    predictor_raster_path = ev_layer,
      max_distance_km = 56,
        out_dir = output_path,
          verbose = TRUE, plot=T,
            xlim = c(90, 179.975),
              ylim = c(-41, 40)
              )
summary(corrected_ob)
}
}
