% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_bin.R
\name{env_bin}
\alias{env_bin}
\title{Create temporally binned predictor layer according to species traits}
\usage{
env_bin(
  sp_trait,
  env_path,
  out_dir,
  var_name,
  lon_min,
  lon_max,
  lat_min,
  lat_max,
  date_min,
  date_max,
  lag_days,
  temporal_bin
)
}
\arguments{
\item{sp_trait}{Character. Species vertical habitat use. One of \code{"pelagic"},
\code{"benthic"}, or \code{"benthopelagic"}.}

\item{env_path}{Character. Path to the downloaded CMEMS NetCDF file.}

\item{out_dir}{Character. Directory where the temporally‑binned raster layers
will be saved.}

\item{var_name}{Character. Name of the variable (refer to CMEMS user manual).}

\item{lon_min}{Numeric. Minimum longitude (western boundary).}

\item{lon_max}{Numeric. Maximum longitude (eastern boundary).}

\item{lat_min}{Numeric. Minimum latitude (southern boundary).}

\item{lat_max}{Numeric. Maximum latitude (northern boundary).}

\item{date_min}{Date (or character convertible to Date). Start date.}

\item{date_max}{Date (or character convertible to Date). End date.}

\item{lag_days}{Numeric. Number of days ocean seasonality lags behind
atmospheric seasonality. Default is \code{76} if missing.}

\item{temporal_bin}{Character. How a year will be split. One of \code{"monthly"},
\code{"quarterly"}, or \code{"seasonal"}.}
}
\value{
No return value. The function writes raster files (ASCII format) to
the directory specified by \code{out_dir}. Files are named according to the
temporal bin (e.g., \code{Jan_thetao_mean.asc}, \code{Q1_thetao_mean.asc},
\code{Spring_thetao_mean.asc}).
}
\description{
Create temporally binned predictor layer according to species traits
}
\examples{
\donttest{
env_bin(
  sp_trait = "pelagic",
  env_path = "F:/whaleshark_sdm/cmems/IndoWPac_temp_400m_y2015_y2024.nc",
  out_dir = "F:/whaleshark_sdm/bin",
  var_name = "thetao_mean",
  lon_min = 100, lon_max = 130,
  lat_min = 15, lat_max = 30,
  date_min = lubridate::ymd(19930101),
  date_max = lubridate::ymd(20221201),
  lag_days = 50,
  temporal_bin = "seasonal"
)
}
}
