% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB_sdm.R
\name{TB_sdm}
\alias{TB_sdm}
\title{Building SDM for each temporal bin and create related figures}
\usage{
TB_sdm(
  env_path,
  prj_path = NULL,
  occ_path,
  eval_occ = NULL,
  res_path,
  species,
  x,
  y,
  colin_var,
  sp_accessible_area,
  pseudoabs_method,
  pres_abs_ratio,
  part,
  algorithm,
  thr,
  ensemble
)
}
\arguments{
\item{env_path}{character. Directory path to the environmental predictors used to calibrate the model.}

\item{prj_path}{character. Directory path to the environmental predictors used to project the distribution
in another temporal or spatial extents. At least two separated folders for different scenarios.}

\item{occ_path}{character. Directory path to occurrence table (tab-delimited text).}

\item{eval_occ}{character. Directory path to independent occurrence data for testing the model.}

\item{res_path}{character. Directory path to store the model results.}

\item{species}{character. Name of the column of species name in the occurrence table.}

\item{x}{character. Name of the column of longitude in the occurrence table.}

\item{y}{character. Name of the column of latitude in the occurrence table.}

\item{colin_var}{character. Method to reduce variable collinearity (default NULL). See ENMTML package.}

\item{sp_accessible_area}{character. Restrict for each species the accessible area (default NULL,
all spatial extent defined by the environmental layer will be used). See ENMTML.}

\item{pseudoabs_method}{character. Pseudo-absence allocation method. See ENMTML documentation for details.}

\item{pres_abs_ratio}{numeric. Presence-Absence ratio (values between 0 and 1).}

\item{part}{character. Partition method for model's validation. Only one method can be chosen.}

\item{algorithm}{character. Algorithm(s) to construct ecological niche models. See ENMTML.}

\item{thr}{character. Threshold(s) used for presence-absence predictions. See ENMTML.}

\item{ensemble}{character. Method(s) used to ensemble different algorithms. See ENMTML.}
}
\value{
A list containing:
\item{var_imp}{Data frame of variable importance across algorithms.}
\item{var_plot}{ggplot2 object of variable importance boxplot.}
\item{res_cur}{Data frame of response curve data (from sdm::rcurve).}
\item{rcur_plot}{List of response curve plots (from sdm::rcurve).}
}
\description{
Building SDM for each temporal bin and create related figures
}
\examples{
\donttest{
# Define paths for environmental predictors, projection, occurrences, and results
d_env <- "F:/whaleshark_sdm/bin/Q1"
d_prj <- "F:/whaleshark_sdm/bin/proj"
d_occ <- "F:/whaleshark_sdm/occ/R_typus_q1.txt"
d_res <- "F:/whaleshark_sdm/result_sdm/Q1"

# Run SDM for the first quarter
sdm_q1 <- TB_sdm(
  env_path = d_env,
  prj_path = d_prj,
  occ_path = d_occ,
  eval_occ = NULL,
  res_path = d_res,
  species = "Species",
  x = "x",
  y = "y",
  colin_var = NULL,
  sp_accessible_area = NULL,
  pseudoabs_method = c(method = "ENV_CONST"),
  pres_abs_ratio = 0.3,
  part = c(method = "BLOCK"),
  algorithm = c("BRT", "SVM", "RDF", "MXD"),
  thr = c(type = "MAX_TSS"),
  ensemble = c(method = "SUP", metric = "TSS")
)
}
}
